fix: Change float to double as in the update of the spec (#44)

+++ include/firebolt/accessibility.h
@@ -35,7 +35,7 @@ struct ClosedCaptionsSettings
 struct VoiceGuidanceSettings
 {
     bool enabled;
-    float rate;
+    double rate;
     bool navigationHints;
 };
