+++ src/discovery_impl.h
@@ -32,7 +32,7 @@ public:
 
     ~DiscoveryImpl() override = default;
 
-    Result<bool> watched(const std::string& entityId, std::optional<float> progress, std::optional<bool> completed,
+    Result<bool> watched(const std::string& entityId, std::optional<double> progress, std::optional<bool> completed,
                          std::optional<std::string> watchedOn, std::optional<AgePolicy> agePolicy) const override;
 
 private:
